Cleaned up dynamic links; added basic login state
