+++ attendio/lib/home.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'login.dart';
 import 'signin_funcs.dart';
+import 'dynamic_link_funcs.dart';
 
 // Landing page for post successful login
 class FirstScreen extends StatelessWidget {
