+++ attendio/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import 'event.dart';
 import 'login.dart';
 
 void main() {
@@ -10,21 +11,4 @@ void main() {
       '/test': (context) => EventScreen(),
     },
   ));
-}
-
-class EventScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Event Screen',
-      home: Scaffold(
-        appBar: AppBar(
-          title: Text('This is an event'),
-        ),
-        body: Center(
-          child: Text('Event #1'),
-        ),
-      ),
-    );
-  }
 }
